# swagger.yaml
openapi: 3.0.3
info:
  title: Super API gerenciadora de filmes
  description: API para gerenciamento de filmes, contendo controle completo de elenco, equipe técnica, pessoas, premiações e mídias associadas.
  contact:
    email: gabriel4321uo@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:8080/v1/senai/locadora
paths:
...
  /filme/:
    get:
      tags:
        - filme
      summary: Rota para obter todos os filmes.
      description: Rota para obter todos os filmes.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      filme:
                        type: array
                        items:
                          $ref: '#/components/schemas/Filme'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
    post:
      tags:
        - filme
      summary: Rota usada para inserir um novo filme.
      description: Rota usada para inserir um novo filme.
      responses:
        '201':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 201
                  response:
                    type: object
                    properties:
                      filme:
                        type: array
                        items:
                          $ref: '#/components/schemas/Filme'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_BAD_REQUEST'
        '415':
          description: Contenty Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_CONTENT_TYPE'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
      requestBody:
        description: JSON contendo os dados do request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
  /filme/:id:
    get:
      tags:
        - filme
      summary: Rota usada para obter um filme filtrando pelo id.
      description: Rota usada para obter um filme filtrando pelo id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id do filme a pesquisar.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      filme:
                        type: array
                        items:
                          $ref: '#/components/schemas/Filme'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
    put:
      tags:
        - filme
      summary: Rota usada para atualizar um filme.
      description: Rota usada para atualizar um filme.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id do filme a atualizar.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      filme:
                        type: array
                        items:
                          $ref: '#/components/schemas/Filme'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_BAD_REQUEST'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        '415':
          description: Contenty Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_CONTENT_TYPE'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
      requestBody:
        description: JSON contendo os dados do request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
    delete:
      tags:
        - filme
      summary: Rota usada para apagar um filme.
      description: Rota usada para apagar um filme.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id do filme a apagar.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SUCCESS_DELETED_ITEM'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
...
  /ocupacoes/:
    get:
      tags:
        - ocupacoes
      summary: Rota usada para obter todas as ocupações.
      description: Rota usada para obter todas as ocupações.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      ocupacoes:
                        type: array
                        items:
                          $ref: '#/components/schemas/ocupacoes'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
    post:
      tags:
        - ocupacoes
      summary: Rota para inserir uma ocupação.
      description: Rota para inserir uma ocupação.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id da ocupação.
      responses:
        '201':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 201
                  response:
                    type: object
                    properties:
                      ocupacoes:
                        type: array
                        items:
                          $ref: '#/components/schemas/ocupacoes'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_BAD_REQUEST'
        '415':
          description: Contenty Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_CONTENT_TYPE'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
      requestBody:
        description: JSON contendo os dados do request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ocupacoes'
  /ocupacoes/:id:
    get:
      tags:
        - ocupacoes
      summary: Rota usada para obter uma ocupação pelo id.
      description: Rota usada para obter uma ocupação pelo id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id da ocupação.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      ocupacoes:
                        type: array
                        items:
                          $ref: '#/components/schemas/ocupacoes'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
    put:
      tags:
        - ocupacoes
      summary: Rota para atualizar uma ocupação.
      description: Rota para atualizar uma ocupação.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Id da ocupação.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_description:
                    type: string
                    example: API para controlar o projeto de Filmes 🎥.
                  development:
                    type: string
                    example: Gabriel
                  version:
                    type: string
                    example: 1.0.0
                  status:
                    type: boolean
                    example: true
                  status_code:
                    type: integer
                    example: 200
                  response:
                    type: object
                    properties:
                      ocupacoes:
                        type: array
                        items:
                          $ref: '#/components/schemas/ocupacoes'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_BAD_REQUEST'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        '415':
          description: Contenty Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_CONTENT_TYPE'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
      requestBody:
        description: JSON contendo os dados do request
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ocupacoes'
    delete:
      tags:
        - ocupacoes
      summary: Rota para deletar uma ocupação.
      description: Rota para deletar uma ocupação.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: id da ocupação a deletar.
      responses:
        '200':
          description: Operação bem sucedida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SUCCESS_DELETED_ITEM'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_NOT_FOUND'
        500 (CONTROLLER):
          description: Internal Server Error CONTROLLER
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_CONTROLLER'
        500 (MODEL):
          description: Internal Server Error MODEL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ERROR_INTERNAL_SERVER_MODEL'
...
components:
  schemas:
    ERROR_BAD_REQUEST:
      type: object
      properties:
        status:
          type: boolean
          description: status false
        status_code:
          type: integer
          format: int64
          description: '400'
        message:
          type: string
          description: Não foi possivel processar a requisição devido a erros de entrada de dados 😣.
        image:
          type: string
          description: https://http.dog/400.jpg
    SUCCESS_DELETED_ITEM:
      type: object
      properties:
        status:
          type: boolean
          description: 'true'
        status_code:
          type: integer
          format: int64
          description: '200'
        message:
          type: string
          description: Item excluido com sucesso 🥳
        image:
          type: string
          description: https://http.dog/200.jpg
    ERROR_INTERNAL_SERVER_MODEL:
      type: object
      properties:
        status:
          type: boolean
          description: 'false'
        status_code:
          type: integer
          format: int64
          description: '500'
        message:
          type: string
          description: Não foi possível processar a requisição devido a um erro interno no servidor. 🤖 [MODEL].
        image:
          type: string
          description: https://http.dog/500.jpg
    ERROR_INTERNAL_SERVER_CONTROLLER:
      type: object
      properties:
        status:
          type: boolean
          description: 'false'
        status_code:
          type: integer
          format: int64
          description: '500'
        message:
          type: string
          description: Não foi possível processar a requisição devido a um erro interno no servidor. 🤖 [CONTROLLER].
        image:
          type: string
          description: https://http.dog/500.jpg
    ERROR_CONTENT_TYPE:
      type: object
      properties:
        status:
          type: boolean
          description: 'false'
        status_code:
          type: integer
          format: int64
          description: '415'
        message:
          type: string
          description: Não foi possível processar a requisição, pois o formato de dados ecaminhado não é suportado pelo servidor, apenas deve ser utilizado JSON. 🤖
        image:
          type: string
          description: https://http.dog/415.jpg
    ERROR_NOT_FOUND:
      type: object
      properties:
        status:
          type: boolean
          description: 'false'
        status_code:
          type: integer
          format: int64
          description: '404'
        message:
          type: string
          description: Não foram encontrados dados para retorno. 🗺️❌
        image:
          type: string
          description: https://http.dog/404.jpg
...
    Filme:
      required:
        - nome
        - sinopse
        - data_lancamento
        - duracao
        - bilheteria
        - id_status
        - id_classificacao
      type: object
      properties:
        nome:
          type: string
          description: Nome do filme a inserir
        sinopse:
          type: string
          description: Sinopse do filme a inserir.
        data_lancamento:
          type: string
          format: date
          description: Data de lançamento do filme a inserir.
        duracao:
          type: string
          format: time
          description: Duração do filme a inserir.
        bilheteria:
          type: number
          description: Bilheteria do filme a inserir.
        valor:
          type: number
          description: Valor do aluguel do filme (caso exista).
        avaliacao_critica:
          type: integer
          format: int64
          description: Avaliação da critica do filme a inserir.
        avaliacao:
          type: integer
          format: int64
          description: Avaliação do publico geral do filme.
        id_status:
          type: integer
          format: int64
          description: id do status do filme.
        id_classificacao:
          type: integer
          format: int64
          description: id da classificação do filme.
...
    ocupacoes:
      required:
        - nome
        - descricao
        - id_setor
      type: object
      properties:
        nome:
          type: string
          description: Nome da ocupação
        descricao:
          type: string
          description: Descrição da ocupação.
        id_setor:
          type: integer
          format: int64
          description: Id do setor que a area pertence.
...